+++ osu.Game/Screens/Select/Carousel/DrawableCarouselBeatmapSet.cs
@@ -229,12 +229,12 @@ namespace osu.Game.Screens.Select.Carousel
 
         public class FilterableGroupedDifficultyIcon : GroupedDifficultyIcon
         {
-            private readonly List<CarouselBeatmap> items;
+            public readonly List<CarouselBeatmap> Items;
 
             public FilterableGroupedDifficultyIcon(List<CarouselBeatmap> items, RulesetInfo ruleset)
                 : base(items.Select(i => i.Beatmap).ToList(), ruleset, Color4.White)
             {
-                this.items = items;
+                Items = items;
 
                 foreach (var item in items)
                     item.Filtered.BindValueChanged(_ => Scheduler.AddOnce(updateFilteredDisplay));
@@ -242,10 +242,16 @@ namespace osu.Game.Screens.Select.Carousel
                 updateFilteredDisplay();
             }
 
+            protected override bool OnClick(ClickEvent e)
+            {
+                Items.First().State.Value = CarouselItemState.Selected;
+                return true;
+            }
+
             private void updateFilteredDisplay()
             {
                 // for now, fade the whole group based on the ratio of hidden items.
-                this.FadeTo(1 - 0.9f * ((float)items.Count(i => i.Filtered.Value) / items.Count), 100);
+                this.FadeTo(1 - 0.9f * ((float)Items.Count(i => i.Filtered.Value) / Items.Count), 100);
             }
         }
     }